requirements.txt
appdirs==1.4.3
Django==1.11.1
django-kronos==1.0
mysqlclient==1.3.10
packaging==16.8
pycrypto==2.6.1
pyparsing==2.2.0
python-crontab==2.2.2
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
